Allow the one sided components to correctly retrieve the op to be applied. Correct the MPI validation process of the MPI_Accumulate arguments.
Fix another potential problem not yet reported. If we convert the
MPI datatypes direclty into OPAL datatypes, we will restrict their
number to the locally different types. Which might not be identical
on the remote node, if we are in a heterogeneous environment. So,
for MPI One sided only deal with MPI level types, never simplify
them on OPAL types (at least on the args). The unfortunate
outcome is that we need to create the args for all datatypes.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@24466 63e3feb5-37d5-0310-a306-e8a459e722fe